requirements.txt
google-generativeai
tavily-python
aiohttp
tenacity
